Fixed a few critical warnings and project configurations.
- Some proto fields were float when they should have been int.
- Proto message generation command needed extra quotes.
- Some projects were missing explicit project level dependencies.
- Change C++ version to 17 for some projects to reduce errors.
- Removed the proto include files from Game.h
- Fixed setMap in the backend, using the std::string variant and getting
rid of the char* variant.

+++ bwapi/include/BWAPI/GameData.h
@@ -26,10 +26,10 @@ namespace BWAPI
 
     std::vector<PlayerID> replayVisionPlayers;
 
-    float latencyTime;
-    float remainingLatencyTime;
-    float elapsedTime;
-    float millisecondsPerFrame;
+    int latencyTime;
+    int remainingLatencyTime;
+    int elapsedTime;
+    int millisecondsPerFrame;
     float averageFPS;
 
     int countdownTimer;
